crée page acceuil

+++ src/main/java/com/example/agendadesevenements/entity/Categorie.java
@@ -8,7 +8,7 @@ public class Categorie {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idCategorie;
+    private Integer idCategorie;
 
     @Column(name="nomCategorie")
     private String nomCategorie;
